Move CommentError and create_comment to CommentHelper
